+++ python/src/scipp/plotting/formatters.py
@@ -1,4 +1,4 @@
-from .._utils import value_to_string
+from ..utils import value_to_string
 from .._scipp.core import to_unit, Unit
 import numpy as np
