+++ python/src/scipp/show.py
@@ -5,7 +5,7 @@ from html import escape
 import numpy as np
 from ._scipp import core as sc
 from . import config
-from ._utils import is_data_array, hex_to_rgb, rgb_to_hex
+from .utils import is_data_array, hex_to_rgb, rgb_to_hex
 from .html import inject_style
 
 # Unit is `em`. This particular value is chosen to avoid a horizontal scroll
