+++ python/src/scipp/utils/__init__.py
