+++ python/src/scipp/utils/collapse_and_slices.py
