+++ python/src/scipp/utils/colors.py
