+++ python/src/scipp/utils/comparison.py
