+++ python/src/scipp/utils/profile.py
