+++ python/src/scipp/utils/to_string.py
