+++ python/src/scipp/utils/typing.py
