+++ python/tests/utils/collapse_and_slices_test.py
